extract up/down samplings to another class
